Jus thinkin. I think that a->n update in lflist remove is needless. If a->n is updated by another node, then all previous transactions will be destined to fail since every node n locks n->p before writing a->n and updating n->p.gen. If n updates a->n, then it couldn't have been unlocked. I think the update is only helpful to prevent needless attempts to swing pat->n after the first swing succeeds.
In fact, that's related to THE BUG, WHICH I FOUND, ALL BY MYSELF. You
swing n but don't update a->n cause you're in list_pop. pat detects it's
been helped and doesn't touch a->n. Now n can no longer be reached from a.

But there's another issue here. In this log, a shouldn't have swung pat,
since patp should have been a. That's just bizarro. Never mind! patp was
locked!

Next bug: if p swings n and a detects it in help_next or help_prev, it'll
correctly recognize that a is dead. But then it'll assume that some other
thread "won" ownership of a and return -1, losing 'a' for good. Fix is:
have everyone attempt to update p. This means that remove always
guarantees that a is off the list, and has become re-addable at least once
after that gen.