Aforementioned changes are actually in. Why did I write that before? Also, I can rewrite things a bit to reduce reads later. Let help_next and help_prev take ptr to n/p.
Why did I need to update gen on addition again? For one thing, the
swinging behavior in help_next depends on it. You want n->p.gen to change
each time pat reoccurs as n->p.

Some issue with help_prev and race between add and remove. When A is being
added pnn == a should be false. But in this case you want the help_prev to
fail. What was the problem?