Should I update n->p first or p->n. If n->p, then gen needs to go in a->n. Removal may help over-eagerly since you thus can't detect whether a has different gen when writing a->p. But it's a rare edge case that only affects perf, as you get an a->n check for free after help_next.
Actually, it could affect correctness. a->n check is OK, but then a is
removed and reinserted and you might remove at wrong gen, and thus from
wrong list. a->p.gen++ and a->p.gen = NULL before every a->n.gen++, and n
is only checked if a is still live. But suppose you read p and n between
a->p.gen++ and the actual write to n->p (so a is live), then a is
reinserted with just a->n.gen++ and you succeed in a spurious lock.

Fix: just increment a->p.gen upon every add, at no extra cost. So p.gen is
adds+locks while n.gen is adds.

Should flinref_up in help_next trigger continue or fail? n has refs till
removal is finished. Removal isn't finished till a->n is written. Unless a
is itself helped, whereupon no one writes a->n. Re-read n for
eternity. But a->n == n doesn't mean that n wasn't reinserted and n->p not
updated. For now, solve this by clearing n after a swing. It seems like
there should be another way.