Thing is, without n.gen, how do you tell whether n->p != a ==> a dead. n := a->n. n removes, n->p != a read, n reinserted, n == a->n, so a falsely "dead". Marking? Could bring back ngen and move a->gen into a 5'th word. BUT I DON'T WANNA. n->p should be incremented if n is readded. n->p can't be dead unless n->p is locked. read n->p, if non-a, read a->n, read n->p again. If n->p is eq, then a->n is still n and a is dead.
BTW, there was something that was easier if you allowed remove to return
without finishing removal. Or rather, that could work somehow. Damn. How?
If you locked a->n rather than a->p, there'd be no need to help nil->n and
a->p would never be locked. So you just swing nil.n->n->p and be done. The
downside is that you need more work to tell a half-add from a half-remove.
a->n->p != a. Do you finish a remove by setting a->n->p = a, or finish an
add by setting a->n = a->n->p? And what about help_prev is list_add? duh.

Anyway, consider scheduler. So you do lflist_enq when switching from a
runnable thread, with the last gen for which you know t to be runnable, so
as to avoid enqueing a DESCHED thread. If the lflist_remove in yield
returns while the preemptor is still popping t in switch_to_electee, then
the add might fail. Could have switch_to_electee attempt the same add. If
the add in lflist_add_rear fails, it's cause switch_to_electee was
removing, and that add will succeed. Same risk of adding a descheduled
thread, if you consider that t might be yielded to, and then desched.