"Fixed" this by not setting n in enq till after p->n is written. I'm essentially back where I started with lflist_del not being able to help lflist_enq. But that's pointless until a completed lflist_del on the right gen can ensure that the next lflist_enq succeeds - which is what I need for the scheduler.
Also may be able to avoid the n->locked case in help_next, and to have
n-aborts skip trying to lock p. Actually, the skip will be necessary to
avoid locking a node that n-aborted because it's still being enqued.

Bug: seeing a linref_account failure. Could be the same flinref_up(nil)
bug I saw before. I added asserts to nalloc to catch it, but no luck?