Fixed: aforementioned change is in. Bug: losing nodes.
Can't del just help enq? If (pn = p->n) != a && a->p == p && p->n == pn,
could it mean anything other than "p->n is being added?" If so, and
help_prev is put first in lflist_del, then it'll perform the first step of
the enq. help_prev is currently second in order to prevent locking a->n if
a->p.gen != a.gen. Save a CAS upon enq at the cost of another a->p read in
del.