Yeah, still seeing bugs. Could be because of p.helped. Right now, seeing a->p->n == nil, which I don't expect to ever happen. I should also be seeing lost anchors since p or n aborts can still lock a->p if a->p has yet to finish an enqueue. Likely that this is somehow propogating rather than just causing lost nodes - since a->p->n hasn't been swung before the lock, and a->n->p might be.
Going to just let lflist_del complete the enq. But it can't blindly write
n->p like in help_next because a might have been helped or deleted by
another thread. So just check whether p->n == a->n? But that's EXACTLY
when you want to swing it to help the enq. Check a->n.locked? That should
only be true when the enq is done, as no one can help a til a->n->p == p,
and a completed lflist_del(a) should have swung a->p->n first.

Also need to avoid helping an enq if it wrote a->p = p but then failed to
write p->n = a and hasn't found a p' to write a->p = p'. Can check p->n ==
a->n, but then what if enq fails because of p->n.gen and then lflist_del
succeeds in helping enq. And then perhaps even finishes deleting a. And
then enq still tries to write p'->n. Could go back to overloading with a
help bit and storing pn.gen, but that didn't work well in the distant
past, nor the recent past. Could just cas on a->p in enq. In fact, with
the new helping scheme, enq can just return after the first cas and let
other operations finish it. Is that more costly? Probably.