You could almost get everything that you need for del-enq shootdown if you just don't clear a->p upon lock(&a->p). But the main issue is that a->n might get overwritten in the enq before nnp fixup is done. Could have everyone do nnp fixup then. But it seems like writing nil.p first in enq is the better way. Suppose you use nil.p.helped. help_prev is "fine" because nil.p->n will be nil if n's being added. The main thing is to distinguish between in-flight del n and in-flight enq n while deleting a | a->n == nil and nil->p == a->nn. a->n.helped won't tell you. n->n.locked || n->n != nil will. Should I implement this now or later?
By the way, is shooting down enq important for the scheduler? It may be
for palloc. Shooting down del is important if a has been yielded to
directly (so still on foreign->runq and then is descheduling on foreign
CPU and its yield(-1) does del with interrupts on. Then the del might be
permanently unreversable. Suppose it had interrupts off. Then the del will
complete and sched_release will redo the failed enq in make_runnable - so
that's OK. Suppose yielding to threads on foreign runqueues deletes them
in order to move them to C, which is defensible anyway. You lose
guarantees about LENGTH of run, but a is still getting a chance to run
before its deadline. You can't (easily) guarantee anything more even if
you keep a on foreign->runq, since it might miss its turn. This all sounds
fine, but I dunno - carefully crafting this ties scheduler policy to
lflist policy. Plus then you have to do the del in yield with interrupts
off for E[cas_loops] = O(NCPUS) loops.

Yeah, doing shootdown isn't worth it now.