Making some dangerous cuts to nnp swing. It's not enough to detect SOME change in n/np over the duration of the lflist_del and use that to rule out helping. n might have re-enqueued itself, after all. You also need an up-to-date picture of a->n. But only p-aborts need to recheck, as a->n will be locked and no helper will unlock it since we've already set p->n = n. If np changes gen AFTER the del is one, then enq(n) completed. pt(pn) != pt(a) says nothing, as pn could have just been helped.
By the way, it's pretty trivial to do cooperating dels now. And now that
there's a case for having p-aborters re-read n, could do another del_won
scheme so that the winner is guaranteed to be the thread that won p->n if
one did. ie - clear a->n.helped upon lock.

Note that you can simply omit the np write and let it get fixed by p. But
that's costlier?