Notes: can't omit np read, as otherwise it might be np from different n and a gen mixup might ensue. onp should be read at the earliest point when you know no new nodes can be added (ie n.locked && p->n = n).
Yeah, if you could ensure that you always get the node with the most
up-to-date picture of n/np, it would save some reads.

You can at least clear p.helped in help_prev, which should only trigger if
enq is done. A bunch of ways to deduce that, actually.