Dang, wrote up a nice log of the bug and emacs lost it. Something like this:
nil.p = p
a.n = nil
a.p = nil .helped=1
d.n = a  .locked=1
d.p = nil
nil.n = a

Just a case of seeing the nnp swing firsthand. I think I really identify
with "a".

a is mid-enq and it set a->p = d, then del(d) was about to swing nil.p=a
but the universe paused itself. So it was just a faulty assert. Added that
case, but maybe I should rethink the overall style tomorrow. Just passed a
large test, but the test_del functionality checks are off.