Still messing around with help_prev loops. It's worrying that the test isn't failing, actually.
By the way:
- Moved *n read out of help_next loop.
- Moved *p read out of help_prev loop. Got rid of spurious rechecks.

You can avoid the a->n recheck on calls to help_next returning 0, since
the lock will do that.

Can you also just not take flinrefs on n until/unless you actually gotta
write n->p?  That means that you may or may not have a ref if you win the
del, depending on helpfulness of help_next. But that's fine. You can
spuriously pass help_next if you read mistyped mem, since there's no n
check on ret 0. But if del(a) wasn't helped, then since pt(a->n) is fixed
after pn is written, the winner must have a ref until it clears a->n and
a->p, forcing faults on recheck. If del(a) *was* helped, then you'll
p-abort. And you'll have to get a ref on p.

It seems like it should be possible to avoid rechecking a->p if
!pn->locked. But tests are failing. Can at least avoid the recheck if
you're going to do a p update to help enq(a | a.nil).

Also, I think the p check in help_next is pointless now that all enques
are properly helped via nnp check, and now that del(a) == -1 if it's
called before the first step in enq. Well, suppose you have an old del(a)
call still going just after failed pn write and a new enq(a) sets n. Then
help_next will swing np before nil.p->n is written.

How's this different from the case now? Why do I do that early
help_prev if I can still get unsafe helping in next without a p check?

Could you reorder help_prev/next? You have to fail the n lock if a->p.gen
!= a.gen. Reordering would just mandate the p recheck, no?

I have to write down guarantees related to locking n := a->n and
help_prev(n). Again, if g == a->n.gen && pn == p->n && g == a->n.gen, and
pn can't be swung nor has it been swung, then because every read of p->n
is followed by an increment of a->n.gen, any other del(a) has an older
picture of p->n and will fail to write it till it rechecks n. And if g >
a->n.gen, any other help_prev(n) will fail to unlock a->n.

The thing is that this conflates reads of p->n with meaningful updates to
a->n. It would be nice to not recheck n->p when n is unchanged. Could do
something like thresholds. If a->n.gen - n.gen >= N, recheck n->p. And
update gen by N in enq.

Subtlety about n->locked check in help_next. I've KINDA been making
deductions from the fact that !n->locked means that no thread has started
del(a). But that's not true - could be a cleared lock from
help_prev(n). Cleared lock behaves like a restart, it seems. In this case
!n->locked means that no thread has finished del.