Working on flinrefs and progress testing.
I think flinref_up(n) == 0 || a->n != n, and the same for p. a->n and a->p
point to undeleted anchors until del(a) has a winner. The winner of del(a)
clears p and n before dropping refs.

How about pp? True for a, true for pp. But you'll have to recheck p first
anyway.