It's was the adderall. Looks good, but still test_del still fails because it assumes that del(a) will shoot down enq(a). Should think about whether I'll need to support that. But right now I'm going to do fully cooperating dels.
After the win, the only thing that needs doing is fixing up np and
nnp. And fixing up np is optional. And then casx(a->n) and casx(a->p)
rather than blind writes. Best bet is to find a better win criteria, so
that only n aborts or pn_ok will win, and so the winner always has the
most recent picture of np. Could do other things with a->n.helped to
detect this.

But even before that, let's try clearing a->p.helped after enq is done.