Simple a->p update wasn't working because I'd been assuming that del-winner's final np update doesn't need to happen. But it does, or else help_prev(n) fails to read pp after it gets cleared. The endgame is to not clear a->p, but it's still better to do an updx_modhelped.
By the way, that thing before about win criteria was bad. p aborts obv
have to be possible winners, as there could be only one del(a) which
locked a->n and then p-aborted. Apart from making it more likely that the
winner has pn_ok if anyone does, there's nothing to improve on this front.