No, legit bug. If help_prev is first, then it might clear a->p.helped before help_next fixes up nil.p. If help_next is first, then it might fix nil.p before nil->p->n is set to a.
What if you just keep a->n nilled until enq(a) is done? That'll work, but
it gets harder to assert things. Or you could go all in and let del finish
up enq. The only nice things are the most aggressive.

I think nilling is the better option.

Again, why didn't this come up earlier?

And could you use *on to avoid the a->p.gen recheckin help_next? Except
for the first iteration, *on must have had on->p == a and then a->p.gen
was a.gen. If it doesn't any more... well, it was reinserted.

BUT! If a isn't being deleted, then !a->n.locked and !a->n.helped. If a
isn't being deleted and a->n->p != a, then a must be mid-enq. So you can
NOT recheck gen but simply check for n->helped || n->locked. If you pass
but a->p.gen != a.gen, then you help another delete, and that other delete
must have another winner, since the winner is the one who first writes n
when !n->helped and !n->locked. You won't do that since help_prev will
fail when reading a->p.gen. Well, it's even more subtle:

What if you call del(a) before enq(a) writes p->n, so that help_prev(a)
finds a->p.gen is correct? I was about to say impossible, because then
!n.helped and !n.locked. But if you call del(a) with an earlier gen and it
finds n.helped || n.locked, then another del(a) might finish and enq(a)
start and then help_prev finds the right gen. So what? The danger is an
early swing of nil.p, before nil->p.n is written. The old del call must
fail to lock n since n has been cleared before the enq, and the next
help_next call will fail if nil.p is bad. Maybe you don't even need to
mention this to prove it.

Does this fix the bug? You could pass an early help_next as described
above, and then help_prev is again clearing p.helped for a node being
enqueued. The only thing to do for that case is force help_next to fail by
nilling a->n. Or if you nil a->p? That does nothing, as realp := nil->p->n
will have realp->n == a and will need to un-nil it to delete. Best not go
there.