Did what I said I would. By the way, things were failing weirdly because I'd deleted the early help_prev call even before figuring everything out.
Because help_prev assumes pn is recent if p is recent, it'll fail if you
preread p but not pn. Easy enough to find and fix.

Actually, this fancy help_next deduction fails in the simple case where n
is mid-deletion and hasn't set n-n->p = a. Ouch.

Either detect that or back to square one. What if you just precheck that
pn == a before entering help_next? That'd work, but then you still need to
avoid writing n->p when a->p.gen is wrong. Can go back to previous scheme,
but...

What if you didn't set n.helped when you updated pn in help_prev, but
rather set it in del(a) pn write and help_prev ppn write and preserved it
from there? Then help_next can indeed detect whether np != a because of
interrupted del or because of incomplete enq. And it might be possible to
leverage that later for enq shootdown. As it is, I've already remarked
that n.helped is pretty useless with the new win scheme.

Gonna try re-read p scheme to sanity check this and then I might do the
above.